+++ Importer/CustomerImporter.php
@@ -14,6 +14,8 @@ use App\Customer\CustomerService;
 use App\Entity\Customer;
 use App\Validator\ValidationFailedException;
 use KimaiPlugin\ImportBundle\Model\ImportData;
+use KimaiPlugin\ImportBundle\Model\ImportModel;
+use KimaiPlugin\ImportBundle\Model\ImportModelInterface;
 use KimaiPlugin\ImportBundle\Model\ImportRow;
 use Symfony\Component\Validator\Validator\ValidatorInterface;
 
@@ -46,13 +48,16 @@ final class CustomerImporter implements ImporterInterface
     }
 
     /**
+     * @param ImportModel $model
      * @param array<ImportRow> $rows
-     * @param bool $dryRun
      * @return ImportData
      */
-    public function import(array $rows, bool $dryRun): ImportData
+    public function import(ImportModelInterface $model, array $rows): ImportData
     {
+        $dryRun = $model->isPreview();
         $data = new ImportData('customers', array_keys($rows[0]->getData()));
+
+        $createdCustomers = 0;
 
         foreach ($rows as $row) {
             try {
@@ -61,6 +66,7 @@ final class CustomerImporter implements ImporterInterface
                 if (!$dryRun) {
                     $this->customerService->saveNewCustomer($customer);
                 }
+                $createdCustomers++;
             } catch (ImportException $exception) {
                 $row->addError($exception->getMessage());
             } catch (ValidationFailedException $exception) {
@@ -70,6 +76,10 @@ final class CustomerImporter implements ImporterInterface
             }
 
             $data->addRow($row);
+        }
+
+        if ($createdCustomers > 0) {
+            $data->addStatus(sprintf('created %s customers', $createdCustomers));
         }
 
         return $data;